25-02-2023 15:16 nuevo

+++ app/src/main/java/com/example/invernadero/ActivityCrearRegistro.java
@@ -158,17 +158,17 @@ public class ActivityCrearRegistro extends AppCompatActivity {
         return ThreadLocalRandom.current().nextInt(minimo, maximo + 1);
     }
 
-    public double numeroTemperatura() {
+    public Integer numeroTemperatura() {
         int min_val = 10;
         int max_val = 50;
-        double randomNum = Math.random() * (max_val - min_val);
+        int randomNum = (int) (Math.random() * (max_val - min_val));
         return randomNum;
     }
 
-    public double numeroHumedad() {
+    public Integer numeroHumedad() {
         int min_val = 10;
         int max_val = 50;
-        double randomNum = Math.random() * (max_val - min_val);
+        int randomNum = (int) (Math.random() * (max_val - min_val));
         return randomNum;
     }
 }
